requirements.txt
portkey-ai==1.2.2
python-dotenv==1.0.0 